% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{apply_threshold}
\alias{apply_threshold}
\title{Apply bootstrap threshhold}
\usage{
apply_threshold(bootstrap_res, true_diff)
}
\arguments{
\item{bootstrap_res}{output of [corr_bootstrap_diff()]}

\item{true_diff}{output of [corr_diff()]}
}
\value{
a matrix
}
\description{
Apply bootstrap threshhold
}
\examples{
\dontrun{
apply_threshold(bootstrap_res, true_diff)
}
}
