% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PotentialLinks.R
\name{assign.tl.category}
\alias{assign.tl.category}
\title{Assign trophic level category}
\usage{
assign.tl.category(tl.a, tl.b)
}
\arguments{
\item{tl.a}{trophic level of first species}

\item{tl.b}{trophic level of second species}
}
\value{
character with potential values "same", "adjacent", or "other". Other represents species separated by >= 1 trophic level
}
\description{
Assign trophic level categorization of a pairwise interaction
}
\examples{
assign.tl.category(1,3)

}
