% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignTrophicLevel.R
\name{AssignTrophicLevel}
\alias{AssignTrophicLevel}
\title{Assign trophic levels and abundances to a set of species}
\usage{
AssignTrophicLevel(num.sp, trophic.levels,
  abundance.distribution = "power-law", scaling.law.tl = F,
  scaling.exponent.tl = 0, basal.abundance = 1000, ...)
}
\arguments{
\item{num.sp}{number of species}

\item{trophic.levels}{number of integer trophic levels in the community}

\item{abundance.distribution}{S.A.D. of the community, either "uniform", "power-law", "lognormal" or "gambin"}

\item{scaling.law.tl}{whether abundances scale with trophic level following a power law}

\item{scaling.exponent.tl}{exponent of the power law trophic level abundance scaling}

\item{basal.abundance}{abundance of the basal trophic level}

\item{...}{additional parameters to specify the S.A.D. (see \code{\link{GenerateProbNumbers}})}
}
\value{
dataframe containing species integer ID, abundances and trophic levels
}
\description{
Assigns a trophic level and abundance to a given number of species, constrained by ecologicaly plausible assumptions
}
\examples{
abundance.list <- AssignTrophicLevel(num.sp = 20, 
trophic.levels = 4, abundance.distribution = "gambin", scaling.law.tl = TRUE, scaling.exponent.tl = 0.75, 
basal.abundance = 1000, gambin.alpha = 2, gambin.maxoctave = 8)

}
\author{
David García-Callejas, \email{david.garcia.callejas@gmail.com}
}
