% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{facetAdjust}
\alias{facetAdjust}
\title{Adjust x-axis on facet_wrap}
\usage{
facetAdjust(x, pos = c("up", "down"), newpage = is.null(vp), vp = NULL)
}
\arguments{
\item{x}{plot object}

\item{pos}{up or down}

\item{newpage}{}

\item{vp}{}
}
\value{
graphical output from function grid.draw
}
\description{
this function (from http://stackoverflow.com/questions/13297155/add-floating-axis-labels-in-facet-wrap-plot?lq=1)
adjusts the x-axis on facet_wrap with uneven number of columns
}
