% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{fit_model}
\alias{fit_model}
\title{Fit a regression model to gene expression}
\usage{
fit_model(expr, donor_vec, engine)
}
\arguments{
\item{expr}{column of gene expression}

\item{donor_vec}{donor vector}

\item{engine}{engine to ues}
}
\value{
a vector of residuals
}
\description{
Fit a regression model to gene expression
}
\examples{
\dontrun{
fit_model(gene, donors, "blmer")
}
}
