% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateLabelDf.R
\name{GenerateLabelDf}
\alias{GenerateLabelDf}
\title{Generate labels for Tukey test analyses}
\usage{
GenerateLabelDf(my.data, HSD, flev, vspace = 0.5)
}
\arguments{
\item{my.data}{dataframe from which anova and Tukey tests were performed}

\item{HSD}{Tukey test}

\item{flev}{level over which to generate labels}

\item{vspace}{vertical spacing above the boxplot}
}
\value{
dataframe with labels for each level, suitable for using as geom_text in ggplot2 plots
}
\description{
For a proper example, see http://stackoverflow.com/questions/18771516/is-there-a-function-to-add-aov-post-hoc-testing-results-to-ggplot2-boxplot
}
