% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneratePerturbation.R
\name{GeneratePerturbation}
\alias{GeneratePerturbation}
\title{GeneratePerturbation}
\usage{
GeneratePerturbation(perturbation.type = "white", timesteps = 10,
  perturbation.mean = 0, perturbation.sd = 1, perturbation.prob = NULL,
  perturbation.max = NULL, perturbation.increment = NULL)
}
\arguments{
\item{perturbation.type}{values: "white", "pulse", "incremental"}

\item{timesteps}{length of the time series}

\item{perturbation.mean}{mean value}

\item{perturbation.sd}{standard deviation}

\item{perturbation.prob}{for perturbation type = pulse, probability of observing a perturbation}

\item{perturbation.max}{for perturbation type = incremental, maximum value for the perturbation. If 0, no maximum}

\item{perturbation.increment}{for perturbation type = incremental, increment in MEAN at each timestep}
}
\value{
numeric vector of given length
}
\description{
Generate a perturbation of a certain length in a time series, given by values in a vector
}
\examples{
GeneratePerturbation(perturbation.type = "pulse",timesteps = 100,perturbation.mean = 1,perturbation.sd = 3,perturbation.prob = 0.2)
}
