% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{get_empirical_p}
\alias{get_empirical_p}
\title{Get the empirical p-value for each link}
\usage{
get_empirical_p(dir)
}
\arguments{
\item{dir}{directory where permutations are stored}
}
\description{
Get the empirical p-value for each link
}
\examples{
\dontrun{
get_empirical_p(tempdir())
}
}
