% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCommunitySize.R
\name{GetCommunitySize}
\alias{GetCommunitySize}
\title{Get the size of a community in each cell of a grid}
\usage{
GetCommunitySize(spatial.grid, plant.population)
}
\arguments{
\item{spatial.grid}{the spatial grid containing the cells. See grid_structure.R}

\item{plant.population}{nested list with the plant population of each cell. See model.R}
}
\value{
vector of size equal to the number of cells, with the size of the community at each cell
}
\description{
By community size I mean here the net size of all individuals
regardless of their species identity. 
For example, five individuals of five different species and size "1" each
will give a community size of "5"; the same number will be the result
of having only one individual of size "5"
}
