% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetIndex.R
\name{GetIndex}
\alias{GetIndex}
\title{dplyr equivalent to which}
\usage{
GetIndex(data, conditions)
}
\arguments{
\item{data}{dataframe}

\item{conditions}{logical vector}
}
\value{
numeric vector of indices satisfying conditions
}
\description{
dplyr equivalent to which
}
