% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInteractionsProb.R
\name{GetInteractionsProb}
\alias{GetInteractionsProb}
\title{Return probabilities of occurrence of different interaction types across trophic levels,
based either on predator-prey topologies (cascade and niche), the literature review
by García-Callejas et al. (2017) or random assignment of probabilities}
\usage{
GetInteractionsProb(antagonistic = "niche", other.int = "random")
}
\arguments{
\item{antagonistic}{topology of antagonistic interactions, either "cascade","niche" or "empirical"}

\item{other.int}{topology of other interaction types, either "random" or "empirical"}
}
\value{
data.frame interaction.probs with columns "interaction.type", "tl.category" and "prob"
}
\description{
Return probabilities of occurrence of different interaction types across trophic levels,
based either on predator-prey topologies (cascade and niche), the literature review
by García-Callejas et al. (2017) or random assignment of probabilities
}
\examples{
GetInteractionsProb("empirical","empirical")
}
