% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{grid_arrange_shared_legend}
\alias{grid_arrange_shared_legend}
\title{draw several plots with only one common legend}
\usage{
grid_arrange_shared_legend(..., ncol = length(list(...)), nrow = 1,
  position = c("bottom", "right"))
}
\arguments{
\item{...}{ggplot2 objects}

\item{ncol}{number of columns}

\item{nrow}{number of rows}

\item{position}{position of the legend}
}
\value{
graphical output from function grid.draw
}
\description{
draw several plots with only one common legend
}
