% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsa.R
\name{gsa}
\alias{gsa}
\title{gene-set enrichment analysis}
\usage{
gsa(geneset, pathways, backg, set_name = "set1", overlap_threshold = 10)
}
\arguments{
\item{geneset}{a character vector of gene ids ENSGID}

\item{pathways}{a data.frame with columns ensgid, go, pathway}

\item{backg}{a character vector of gene ids (ENSGID)}

\item{set_name}{a character string to name the gene set}

\item{overlap_threshold}{an integer to filter pathways with overlap size greater than or equal to this value}
}
\value{
a [dplyr::tibble()]
}
\description{
gene-set enrichment analysis
}
\examples{
\dontrun{
gsa(geneset, pathways, backg)
}
}
