% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactMatrix.R
\name{ImpactMatrix}
\alias{ImpactMatrix}
\title{Calculate impact matrices from abundance data}
\usage{
ImpactMatrix(abundance.data, sign.matrix, a.param = c(0.01, 0.01, 0.001, 0.01,
  0.01), x0 = 1, threshold = 0.001, scale.factor = c(0.1, 0.1, 0.5, 0.1,
  0.1), metric = "impact", return.type = "dataframe", my.model = NULL,
  param.list = NULL)
}
\arguments{
\item{abundance.data}{abundances vector}

\item{sign.matrix}{sign matrix}

\item{a.param}{vector of 5 parameters adjusting the net interaction frequency for each type: 
in the following order: competition, amensalism, antagonism, mutualism, commensalism}

\item{x0}{parameter for the logistic function. In principle, 1 works fine for all simulations}

\item{threshold}{extinction threshold for abundance}

\item{scale.factor}{factor scaling the strength of each interaction type
e.g. a scaling of 1 indicates that the maximum effect
of one sp. over another is of the same order of magnitude
than the population growth rate. Same order as a.param}

\item{metric}{which metric is returned? either "impact" (the net impact), 
"community" (jacobian/community matrix) or "both"}

\item{return.type}{either matrix or dataframe}

\item{my.model}{dynamic equations model for calculating the jacobian}

\item{param.list}{list of parameters for the dynamic model}
}
\value{
either a dataframe or a list with two matrices
}
\description{
Calculate impact matrices from abundance data
}
