% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionFrequency.R
\name{InteractionFrequency}
\alias{InteractionFrequency}
\title{Interaction Frequency of a pairwise interaction}
\usage{
InteractionFrequency(n1, n2, a, x0, threshold)
}
\arguments{
\item{n1}{abundance of the first species}

\item{n2}{abundance of the second species}

\item{a}{parameter of the logistic function. A higher a implies faster approach to the asymptote}

\item{x0}{parameter of the logistic function}

\item{threshold}{threshold below which an abundance is considered zero}
}
\value{
value, positive and non-integer, of interaction frequency
}
\description{
Interaction frequency calculated from net abundances, as a logistic function
with an asymptote at the lowest abundance of the two
}
\examples{
InteractionFrequency(n1 = 10, n2 = 100, a = 0.1, x0 = 1, threshold = 1e-5)
InteractionFrequency(n1 = 10, n2 = 100, a = 0.01, x0 = 1, threshold = 1e-5)
}
