% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteractionStrength.R
\name{InteractionStrength}
\alias{InteractionStrength}
\title{Interaction strength of a pairwise interaction}
\usage{
InteractionStrength(n1, n2, sign.matrix, sp1, sp2, scale.factor)
}
\arguments{
\item{n1}{Abundance of the first species}

\item{n2}{Abundance of the second species}

\item{sign.matrix}{sign matrix of the community}

\item{sp1}{index of the first species in the sign matrix}

\item{sp2}{index of the second species in the sign matrix}

\item{scale.factor}{numeric vector with five elements, giving the maximum strength relative to
species' intrinsic growth rate. If e.g. the scale factor equals 1, a certain interaction
can have a strength of the same order of magnitude that the r parameter}
}
\value{
dimensionless interaction strength value
}
\description{
Interaction strength calculated as the sign of the interaction multiplied by a scaling factor
}
\examples{
InteractionStrength(n1 = 10,n2 = 100,
sign.matrix = matrix(data = c(0,-1,1,0),nrow = 2),sp1 = 1,sp2 = 2,scale.factor = rep(0.1,5))
}
