% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinMap.R
\name{LinMap}
\alias{LinMap}
\title{Remap a vector to a determined range}
\usage{
LinMap(x, from, to)
}
\arguments{
\item{x}{vector to map}

\item{from}{starting point of new range}

\item{to}{end point of new range}
}
\value{
vector of length(x)
}
\description{
Remap a vector to a determined range
}
