% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxLinks.R
\name{MaxLinks}
\alias{MaxLinks}
\title{Maximum number of links for every interaction type}
\usage{
MaxLinks(tl.results, potential.links)
}
\arguments{
\item{tl.results}{a dataframe as given by \code{AssignTrophicLevel}}

\item{potential.links}{a dataframe as given by \code{PotentialLinks(...,return.type = "positions")}}
}
\value{
numeric vector of length 5
}
\description{
Returns the maximum number of links that can be realized in a given community
per interaction type, i.e. if an interaction type was allowed to fill the entire link space
}
\examples{
tl.results <- AssignTrophicLevel(num.sp = 20,
                                 trophic.levels = 4,
                                 abundance.distribution = "gambin",
                                 scaling.law.tl = TRUE,
                                 scaling.exponent.tl = 0.75,
                                 basal.abundance = 1000,
                                 gambin.alpha = 2,
                                 gambin.maxoctave = 8)

within.type.prob <- data.frame(interaction.type = c(rep("competition",3),rep("amensalism",3),rep("antagonism",3),rep("mutualism",3),rep("commensalism",3)),
                   tl.category = rep(c("same","adjacent","other"),5),
                   prob = rep(0.33,15))
within.type.prob$prob[within.type.prob$interaction.type == "antagonism" & within.type.prob$tl.category == "same"] <- 0
within.type.prob$prob[within.type.prob$interaction.type == "antagonism" & within.type.prob$tl.category == "adjacent"] <- 0.75
within.type.prob$prob[within.type.prob$interaction.type == "antagonism" & within.type.prob$tl.category == "other"] <- 0.25
potential.links <- PotentialLinks(tl.results,within.type.prob)
max.links <- MaxLinks(tl.results,potential.links)
}
