% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mode.R
\name{Mode}
\alias{Mode}
\title{Function for calculating the mode of a vector}
\usage{
Mode(x)
}
\arguments{
\item{x}{vector af any length and type}
}
\value{
Most repeated value in the data. In case of ties, the first value is reported
}
\description{
Function for calculating the mode of a vector
}
\examples{
Mode(c("a","b","a","b"))

}
