% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PairwiseCorrelation.R
\name{PairwiseCorrelation}
\alias{PairwiseCorrelation}
\title{Computes the correlation of a pairwise interaction strength matrix or dataframe
following the formulation of Tang et al. (2014)}
\usage{
PairwiseCorrelation(interaction.matrix)
}
\arguments{
\item{interaction.matrix}{NxN matrix with interaction effects of columns on rows}
}
\value{
correlation value
}
\description{
Computes the correlation of a pairwise interaction strength matrix or dataframe
following the formulation of Tang et al. (2014)
}
\examples{
interaction.matrix <- matrix(c(1,2,3,1,1,1,0.4,-2,-5),nrow = 3)
PairwiseCorrelation(interaction.matrix)
}
