% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{palette.values}
\alias{palette.values}
\title{Colors I usually use}
\usage{
palette.values()
}
\value{
six common-used colors from a color-blind palette
}
\description{
Colors I usually use
}
