% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlantLogistic.R
\name{PlantLogistic}
\alias{PlantLogistic}
\title{Logistic function for plant responses}
\usage{
PlantLogistic(x, a, b, c)
}
\arguments{
\item{x}{response}

\item{a}{first parameter}

\item{b}{second parameter}

\item{c}{third parameter}
}
\value{
logistic function
}
\description{
Different parameterizations may be used for plant survival or growth responses
to an arbitrary stress factor. Modelled following fig. 1 of Maestre et al. 2009
}
\details{
parameter space for survival probability:
a = 1
b approaching 0
0.1 < c < 0.98

parameter space for growth:
a = 1
a combination of b approaching 0 and c = 0.9 as a limit; 
b > 0.7 and c > 0.95 as the other limit
}
