% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateProbNumbers.R
\name{plfunc.2}
\alias{plfunc.2}
\title{Power law sampling}
\usage{
plfunc.2(x, c, k)
}
\arguments{
\item{x}{base}

\item{c}{parameter}

\item{k}{exponent}
}
\value{
value of (x*c)^k
}
\description{
Power law sampling
}
\examples{
plfunc.2(10,1,0.75)
}
