% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{read_data}
\alias{read_data}
\title{Read seurat data into dGC format}
\usage{
read_data(path, type = c("seurat", "h5ad", "loom"))
}
\arguments{
\item{path}{filepath}

\item{type}{filetype}
}
\value{
a list with three elements:
}
\description{
Read seurat data into dGC format
}
\examples{
\dontrun{
read_data("path/to/sr.rds")
}
}
