% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{run_dgc}
\alias{run_dgc}
\title{Run differential gene-correlation analysis
results are saved in `dir`, which defaults to the working directory}
\usage{
run_dgc(
  obj,
  split_by = "condition",
  n_iter = 100,
  fit_models = c("none", "blmer", "lmer", "glmer"),
  method = c("pearson", "spearman"),
  formula = stats::as.formula("expr ~ 1 + (1|donor)"),
  min_cells_per_donor = 10,
  n_ctrl = 10,
  n_case = 10,
  replace = FALSE,
  dir = NULL
)
}
\arguments{
\item{obj}{a list created by `prep_cluster_counts()`}

\item{split_by}{column in `obj$obs` to split samples by condition}

\item{n_iter}{number of permutations}

\item{fit_models}{which models to fit}

\item{method}{correlation method}

\item{formula}{formula for mixed effect model}

\item{min_cells_per_donor}{Minimum number of cells per donor}

\item{n_ctrl}{number of control donors to sample}

\item{n_case}{number of case donors to sample}

\item{replace}{sample with replacement}

\item{dir}{output directory}
}
\description{
Run differential gene-correlation analysis
results are saved in `dir`, which defaults to the working directory
}
\examples{
\dontrun{
run_dgc(object, split_by = "disease", n_iter=10, dir = tempdir())
}
}
