% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{run_real_diff}
\alias{run_real_diff}
\title{Calculate the true correlation difference between conditions}
\usage{
run_real_diff(dir, ncores = 1)
}
\arguments{
\item{dir}{directory where [setup_dgc()] has created data}

\item{ncores}{number of cores for residualizing}
}
\description{
Calculate the true correlation difference between conditions
}
\examples{
\dontrun{
run_real_diff(tempdir(), ncores=3)
}
}
