% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{scale_colour_Publication}
\alias{scale_colour_Publication}
\title{Auxiliary color scale for theme_Publication}
\usage{
scale_colour_Publication(...)
}
\arguments{
\item{...}{}
}
\value{
a ggplot discrete scale
}
\description{
Auxiliary color scale for theme_Publication
}
