% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{scale_fill_Publication}
\alias{scale_fill_Publication}
\title{Auxiliary fill scale for theme_Publication}
\usage{
scale_fill_Publication(...)
}
\arguments{
\item{...}{}
}
\value{
a ggplot discrete scale
}
\description{
Auxiliary fill scale for theme_Publication
}
