% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuxiliaryPlotFunctions.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Theme for publication}
\usage{
theme_Publication(base_size = 12, base_family = "helvetica")
}
\arguments{
\item{base_size}{font size}

\item{base_family}{font family}
}
\value{
apply the theme to a ggplot object
}
\description{
Theme for publication
}
