% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validation.R
\name{validate_data}
\alias{validate_data}
\title{validate data format}
\usage{
validate_data(data)
}
\arguments{
\item{data}{input list}
}
\value{
a list
}
\description{
validate data format
}
\examples{
\dontrun{
data <- validate_data(data)
}
}
