% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WithinTypeProb.R
\docType{data}
\name{WithinTypeProb}
\alias{WithinTypeProb}
\title{Probabilities of occurrence of interaction types across trophic levels}
\format{dataframe}
\usage{
data(WithinTypeProb)
}
\description{
Data from a literature review (Garcia-Callejas et al. in prep)
}
\keyword{datasets}
