% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{compute_residuals}
\alias{compute_residuals}
\title{compute residuals of a single-cell data matrix}
\usage{
compute_residuals(
  matrix,
  engine = c("blmer", "lmer", "glmer"),
  formula = stats::as.formula("expr ~ 1 + (1|donor)"),
  cells = NULL,
  donor_vec,
  ncores = 1
)
}
\arguments{
\item{matrix}{count matrix}

\item{engine}{method to fit the model, one of "blmer", "lmer", or "glmer"}

\item{formula}{a formula to use for the model, default is "expr ~ 1 + (1|donor)"}

\item{cells}{a vector of cell identifiers to use, if NULL all cells are used}

\item{donor_vec}{a vector of donor identifiers, must be the same length as cells}

\item{ncores}{number of cores to use for parallel processing, default is 1}
}
\value{
a list()
}
\description{
compute residuals of a single-cell data matrix
}
\examples{
\dontrun{
compute_residuals(count_matrix)
}
}
