% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{corr_diff}
\alias{corr_diff}
\title{Calculate the difference in Gene-Gene correlation across two conditions}
\usage{
corr_diff(
  M,
  obs_1,
  obs_2,
  fit_models = c("none", "blmer", "lmer", "glmer"),
  method = c("pearson", "spearman"),
  ncores = 1
)
}
\arguments{
\item{M}{a matrix of gene expression values, with genes as columns and cells as rows}

\item{obs_1}{a data frame containing cell identifiers and donor information for the first condition}

\item{obs_2}{a data frame containing cell identifiers and donor information for the second condition}

\item{fit_models}{a character vector indicating whether to fit a linear mixed model to the data, one of "none", "blmer", "lmer", or "glmer"}

\item{method}{a character string indicating the method to use for correlation calculation, one of "pearson" or "spearman"}

\item{ncores}{number of cores to use for parallel processing, default is 1}
}
\value{
a matrix of the difference in correlation between the two conditions
}
\description{
Calculate the difference in Gene-Gene correlation across two conditions
}
\examples{
\dontrun{
corr_diff(M, obs_1, obs_2, fit_models = "blmer", method = "pearson", ncores = 4)
}
}
