% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlations.R
\name{corr_diff}
\alias{corr_diff}
\title{Calculate the difference in correlation between two conditions}
\usage{
corr_diff(M, obs_df, by = "status", method = c("pearson", "spearman"))
}
\arguments{
\item{M}{matrix (cells x genes)}

\item{obs_df}{observation data frame with a column for conditions}

\item{by}{column name in obs_df to split by (default: "status")}

\item{method}{correlation method (default: "pearson")}
}
\value{
a matrix of differential correlations
}
\description{
Calculate the difference in correlation between two conditions
}
\examples{
\dontrun{
corr_diff(M, obs_df, by = "status", method = "spearman")
}
}
