% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{corr_permute}
\alias{corr_permute}
\title{Compute permutations of differential gene correlation}
\usage{
corr_permute(
  obj,
  n_iter = 100,
  residualise = FALSE,
  method = c("pearson", "spearman"),
  n_ctrl = 10,
  n_case = 10,
  min_cells = 20,
  replace = FALSE,
  ncores = 1
)
}
\arguments{
\item{obj}{a list in the format of [read_data()]}

\item{n_iter}{number of iterations to perform}

\item{residualise}{whether to fit a linear-mixed-model to remove donor effects from each gene}

\item{method}{method for correlation, pearson or spearman}

\item{n_ctrl}{number of donors in "control" correlation group}

\item{n_case}{number of donors in "case" correlation group}

\item{min_cells}{minimum number of cells per donor to include in the analysis}

\item{replace}{whether to sample donors with replacement}

\item{ncores}{number of cores to use for parallel processing}
}
\value{
a list of matrices, each matrix is the difference in correlation between the case and control groups for each iteration
}
\description{
Compute permutations of differential gene correlation
}
\examples{
\dontrun{
permuts <- corr_permute(obj)
}

}
