% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{mask_from_perm}
\alias{mask_from_perm}
\title{calculate a mask from a list of permutations}
\usage{
mask_from_perm(P, R, alpha = 0.01)
}
\arguments{
\item{P}{list of permutations, from [corr_permute()]}

\item{R}{a matrix of correlations, e.g. the output from [corr_diff()]}
}
\value{
a matrix of the same dimensions as R, with values between 0 and 1 indicating the proportion of permutations that were greater than the observed value in R
}
\description{
calculate a mask from a list of permutations
}
\examples{
\dontrun{
mask_from_perm(perm, real)
}
}
