% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-data.R
\name{prep_cluster_counts}
\alias{prep_cluster_counts}
\title{prep_cluster counts perform basic subsetting prior to running dGC.
cells are filtered to a single cell-type and genes are reduced by considering
the proportion of cells they are expressed in.}
\usage{
prep_cluster_counts(obj, ct, ct_column = "named_celltype", prop_cells = 0.9)
}
\arguments{
\item{obj}{dGA object}

\item{ct}{celltype, "beta cells"}

\item{ct_column}{column of celltypes}

\item{prop_cells}{proportion of cells that need to express a gene to be kept}
}
\value{
a list
}
\description{
prep_cluster counts perform basic subsetting prior to running dGC.
cells are filtered to a single cell-type and genes are reduced by considering
the proportion of cells they are expressed in.
}
\examples{
\dontrun{
prep_cluster_counts(obj, ct = "beta cells", ct_column = "named_celltype", prop_cells = 0.9)
}
}
