% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{run_permutations}
\alias{run_permutations}
\title{Generate permutations}
\usage{
run_permutations(dir, ncores = 1)
}
\arguments{
\item{dir}{directory where [setup_dgc()] has created data}

\item{ncores}{number of cores for residualizing}
}
\description{
Generate permutations
}
\examples{
\dontrun{
run_permutations(100, labels, M, obs)
}

}
