% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{run_permutations}
\alias{run_permutations}
\title{Generate permutations}
\usage{
run_permutations(
  n_iter,
  permutation_labels,
  M,
  obs,
  fit_models,
  formula,
  method
)
}
\arguments{
\item{n_iter}{number of iterations}

\item{permutation_labels}{permutation list}

\item{M}{Matrix}

\item{obs}{observation df}

\item{fit_models}{which models to fit}

\item{formula}{formula for mixed effect model}

\item{method}{correlation method}
}
\description{
Generate permutations
}
\examples{
\dontrun{
run_permutations(100, labels, M, obs)
}

}
