% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation-diff.R
\name{setup_dgc}
\alias{setup_dgc}
\title{Setup the analysis for differential gene-correlation}
\usage{
setup_dgc(
  obj,
  split_by = "condition",
  n_iter = 100,
  fit_models = c("none", "blmer", "lmer", "glmer"),
  method = c("pearson", "spearman"),
  min_cells_per_donor = 10,
  replace = FALSE,
  dir = NULL
)
}
\arguments{
\item{obj}{an object created by [read_data()]}

\item{split_by}{column to split by - the `condition` column}

\item{n_iter}{number of iterations for bootstrap}

\item{fit_models}{regress out donor effects?}

\item{method}{pearson or spearman for correlation?}

\item{min_cells_per_donor}{minimum number of cells per donor}

\item{replace}{allow replacement in permutation?}

\item{dir}{directory to setup analysis in}
}
\description{
Setup the analysis for differential gene-correlation
}
\examples{
\dontrun{
setup_dgc(obj, "case")
}

}
